requirements.txt
aiogrpc==1.8
asttokens==2.4.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
decorator==5.1.1
executing==2.1.0
Flask==3.0.3
googleapis-common-protos==1.65.0
grpcio==1.66.2
idna==3.10
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
numpy==2.1.1
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
protobuf==5.28.2
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyOpenGL==3.1.7
redis==5.1.0
requests==2.32.3
setuptools==75.1.0
six==1.16.0
spotipy==2.24.0
stack-data==0.6.3
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wirepod_vector_sdk==0.8.1
